move create button to its own step, with cost warnings
