+++ src/pages/billing/CreateAzureBillingProjectModal.js
@@ -2,16 +2,15 @@ import _ from 'lodash/fp'
 import { Fragment, useEffect, useState } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import { ButtonPrimary, IdContainer, Link, Select, spinnerOverlay } from 'src/components/common'
-import { icon } from 'src/components/icons'
 import { ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
 import { InfoBox } from 'src/components/PopupTrigger'
 import { Ajax } from 'src/libs/ajax'
-import colors from 'src/libs/colors'
 import { reportErrorAndRethrow } from 'src/libs/error'
 import { formHint, FormLabel } from 'src/libs/forms'
 import { useCancellation } from 'src/libs/react-utils'
 import * as Utils from 'src/libs/utils'
+import { AzureCostWarnings } from 'src/pages/billing/AzureCostWarnings'
 import { validate as validateUuid } from 'uuid'
 import validate from 'validate.js'
 
@@ -170,22 +169,7 @@ const CreateAzureBillingProjectModal = ({ onSuccess, onDismiss, billingProjectNa
           })
         ])
       ])]),
-      div({ style: { paddingTop: '1.0rem', display: 'flex' } },
-        [
-          icon('warning-standard', { size: 16, style: { marginRight: '0.5rem', color: colors.warning() } }),
-          div(['Creating a Terra billing project currently costs about $5 per day. ',
-            h(Link, {
-              href: 'https://support.terra.bio/hc/en-us/articles/12029087819291',
-              ...Utils.newTabLinkProps
-            }, ['Learn more and follow changes.'])])
-        ]
-      ),
-      div({ style: { paddingTop: '1.0rem', display: 'flex' } },
-        [
-          icon('clock', { size: 16, style: { marginRight: '0.5rem' } }),
-          div(['It may take up to 15 minutes for the billing project to be fully created and ready for use.'])
-        ]
-      )
+      h(AzureCostWarnings)
     ]),
     isBusy && spinnerOverlay
   ])
